erhai: keylayout: Rely on Linux key definitions for 0x01 HID usages
We recently introduced two new keys (lock and do not disturb) which
were defined with HID even they were in the 0x01 page; this
introduced a minor bug, as it's overriding the system sleep key,
so that key will now lock the device instead of putting the system
in sleep mode. This CL removes the 0x01 HID definitions and adds
an entry for do not disturb following the Linux definition.

Bug: 409328696
Flag: EXEMPT bugfix
Test: Flashed on device, tested with Logi K650 in Chrome mode
Change-Id: I198caaeddc3b8d0ed58f2fa2dbd659a194e0e5c8